SX Tailwind: Use Tailwind CSS v2.0
The new major version of Tailwind CSS now depends more on configuration. For example the default colors palette was reduced and shifted a little bit, users should define colors explicitly in the config. CSS variables are prefixed with `tw` now and some utilities were renamed.

The main change in SX Tailwind is in the Babel plugin. Styles are now generated and cached based on the configuration so every utility should be configurable. Previously only font family was.

adeira-source-id: 1ad7b55b6bd5a2ddb59914703e302dd076b867ae